requirements.txt
youtube-transcript-api
requests
spacy
python-dotenv
langgraph==1.0.3
openai==1.76.2